perf(showcase): bootstrap the app with angular option strictDi true
Remove client dep angular-ui-codemirror to be able to set strictDi to true.
Move and convert it's code which is an angular directive to rba-codemirror.
Which means, we are now responsible for it. That's ok because it's not that
much code, you couldn't minify it, its injectables weren't annotated and the
last update to the package was in 2015.